dinov2_feature/requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
scikit-learn
opencv-python